iledo/requirements-dev.txt
asgiref==3.2.3
certifi==2019.11.28
cloudinary==1.19.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.2
django-fontawesome==1.0
djangorestframework==3.11.0
mock==3.0.5
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
